+++ lib/chronic.rb
@@ -8,20 +8,21 @@ require 'chronic/time'
 require 'chronic/handler'
 require 'chronic/handlers'
 require 'chronic/mini_date'
-require 'chronic/tag'
 require 'chronic/span'
 require 'chronic/token'
-require 'chronic/grabber'
-require 'chronic/pointer'
-require 'chronic/scalar'
-require 'chronic/ordinal'
-require 'chronic/separator'
-require 'chronic/sign'
-require 'chronic/time_zone'
 require 'chronic/numerizer'
 require 'chronic/season'
 
-require 'chronic/repeater'
+require 'chronic/tag'
+require 'chronic/tags/grabber'
+require 'chronic/tags/ordinal'
+require 'chronic/tags/pointer'
+require 'chronic/tags/scalar'
+require 'chronic/tags/separator'
+require 'chronic/tags/sign'
+require 'chronic/tags/time_zone'
+
+require 'chronic/tags/repeater'
 require 'chronic/repeaters/repeater_year'
 require 'chronic/repeaters/repeater_season'
 require 'chronic/repeaters/repeater_season_name'
