+++ lib/chronic/tags/pointer.rb
