+++ lib/chronic/tags/scalar.rb
