+++ lib/chronic/tags/separator.rb
